nada/requirements.txt
opencv_python==4.5.1.48
numpy==1.20.1
imutils==0.5.4
streamlit==0.57.3
Pillow==8.1.2
